Optimize:optimize SpringValueProcessor
