+++ spring-cloud-starter-tencent-polaris-config/src/test/resources/bean.xml
@@ -11,28 +11,4 @@
 		<property name="name" value="${name:test}"/>
 		<property name="age" value="${age:10}"/>
 	</bean>
-
-	<!--SpringValueDefinitionProcessor to process xml config placeholders	-->
-	<bean class="com.tencent.cloud.polaris.config.spring.property.SpringValueDefinitionProcessor">
-		<constructor-arg index="0" ref="helper"/>
-		<constructor-arg index="1" ref="polarisConfigProperties"/>
-	</bean>
-
-	<!--Placeholder helper functions	-->
-	<bean id="helper" class="com.tencent.cloud.polaris.config.spring.property.PlaceholderHelper"/>
-
-	<!--PolarisConfigProperties	-->
-	<bean id="polarisConfigProperties" class="com.tencent.cloud.polaris.config.config.PolarisConfigProperties">
-		<property name="autoRefresh" value="true"/>
-	</bean>
-
-	<!--springValueRegistry	-->
-	<bean id="springValueRegistry" class="com.tencent.cloud.polaris.config.spring.property.SpringValueRegistry"/>
-
-	<!--Spring value processor of method	-->
-	<bean class="com.tencent.cloud.polaris.config.spring.annotation.SpringValueProcessor">
-		<constructor-arg index="0" ref="helper"/>
-		<constructor-arg index="1" ref="springValueRegistry"/>
-		<constructor-arg index="2" ref="polarisConfigProperties"/>
-	</bean>
 </beans>
